Massive re-organization
This will allow me to more easily integrated it into Discord and begin
the implementation phase.

The core idea is - all hooks and core code goes into
`scroller-utilitiesl` (I don't have a better name for it atm... so we
shall see there).  Anyways, from there - there's a folder that can be
imported for everything, and that becomes the entry point.  When moving
it to an eventual open source library, then we can simply swap out the
import location to a module and it `should just work`...